Split off `HyperHaskell.Interpreter`
